Use size_t for _nettle_write functions.

+++ nettle-write.h
@@ -34,14 +34,14 @@
 
 /* FIXME: Use a macro shortcut to memcpy for native endianness. */
 void
-_nettle_write_be32(unsigned length, uint8_t *dst,
+_nettle_write_be32(size_t length, uint8_t *dst,
 		   uint32_t *src);
 void
-_nettle_write_le32(unsigned length, uint8_t *dst,
+_nettle_write_le32(size_t length, uint8_t *dst,
 		   uint32_t *src);
 
 void
-_nettle_write_le64(unsigned length, uint8_t *dst,
+_nettle_write_le64(size_t length, uint8_t *dst,
 		   uint64_t *src);
 
 #endif /* NETTLE_WRITE_H_INCLUDED */
