+++ write-be32.c
@@ -31,11 +31,11 @@
 #include "macros.h"
 
 void
-_nettle_write_be32(unsigned length, uint8_t *dst,
+_nettle_write_be32(size_t length, uint8_t *dst,
 		   uint32_t *src)
 {
-  unsigned i;
-  unsigned words;
+  size_t i;
+  size_t words;
   unsigned leftover;
   
   words = length / 4;
